#385 Extract ImageLabel to own source file
