+++ src/GUI/popup.cpp
@@ -7,14 +7,12 @@
 #include <QBasicTimer>
 #include <QScreen>
 #include <QApplication>
-#include <QImageReader>
 #include <QVBoxLayout>
-#include <QDesktopServices>
-#include <QFileInfo>
 #if QT_VERSION < QT_VERSION_CHECK(5, 15, 0)
 #include <QDesktopWidget>
 #endif // QT 5.15
 #include "data/imageinfo.h"
+#include "imagelabel.h"
 #include "popup.h"
 
 static QSize thumbnailSize(const ImageInfo &img, int limit)
@@ -31,34 +29,6 @@ static QSize thumbnailSize(const ImageInfo &img, int limit)
 	}
 
 	return QSize(width, height);
-}
-
-class ImageLabel : public QLabel {
-public:
-    ImageLabel(QString path, QSize size, QWidget* parent = 0, Qt::WindowFlags f = Qt::WindowFlags());
-    ~ImageLabel();
-
-protected:
-    void mousePressEvent(QMouseEvent* event);
-
-	QString path;
-
-};
-
-ImageLabel::ImageLabel(QString path, QSize size, QWidget* parent, Qt::WindowFlags f)
-    : QLabel(parent, f), path(path) {
-	QImageReader reader(path);
-	reader.setAutoTransform(true);
-	reader.setScaledSize(size);
-	QImage image = reader.read();
-	QPixmap pixmap = QPixmap::fromImage(image);
-	setPixmap(pixmap);
-}
-
-ImageLabel::~ImageLabel() {}
-
-void ImageLabel::mousePressEvent(QMouseEvent* event) {
-	QDesktopServices::openUrl(QUrl::fromLocalFile(QFileInfo(path).absoluteFilePath()));
 }
 
 class PopupWidget : public QFrame
